connected all js to html pages
