+++ WhatsCooking/src/main/resources/static/css/fragments.css
@@ -6,7 +6,3 @@
 {
     background-color: #368BC1;
 }
-.navbar
-{
-    width: 100%;
-}
