+++ WhatsCooking/src/main/resources/static/css/style.css
@@ -1,0 +1,7 @@
+  .table-responsive{
+    background: #5186ef;
+  }
+
+  .GroceryList_navigation_container{
+    background: #0000000;
+  }
