+++ WhatsCooking/src/main/resources/templates/GroceryList.html
@@ -10,31 +10,15 @@
 
     <script src="https://code.jquery.com/jquery-3.1.1.min.js" integrity="sha256-hVVnYaiADRTO2PzUGmuLJr8BLUSjGIZsDYGmIJLv2b8="
     crossorigin="anonymous"></script>
-
     <!-- Latest compiled CSS -->
     <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.7/css/bootstrap.min.css" integrity="sha384-BVYiiSIFeK1dGmJRAkycuHAHRg32OmUcww7on3RYdg4Va+PmSTsz/K68vbdEjh4u" crossorigin="anonymous"/>
-
     <!-- Theme -->
     <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.7/css/bootstrap-theme.min.css" integrity="sha384-rHyoN1iRsVXV4nD0JutlnGaslCJuC7uwjduW9SVrLvRYooPp2bWYgmgJQIXwl/Sp" crossorigin="anonymous"/>
-
     <!-- CSS Style Sheet File for Entire Web Page -->
-    <link href="../../css/style.css" rel="stylesheet"/>
-
+    <link href="./css/style.css" rel="stylesheet"/>
     <!-- Latest compiled JS -->
     <script src="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.7/js/bootstrap.min.js" integrity="sha384-Tc5IQib027qvyjSMfHjOMaLkfuWVxZxUPnCJA7l2mCWNIpG9mGCD8wGNIcPD7Txa" crossorigin="anonymous"></script>
-
-
-
-  <style>
-  .table-responsive{
-    background: #5186ef;
-  }
-
-  .GroceryList_navigation_container{
-    background: #0000000;
-  }
-  </style>
-
+    <script type="text/javascript" src="./js/GroceryList.js"></script>
   </head>
