+++ WhatsCooking/src/main/resources/templates/fragments.html
@@ -6,6 +6,7 @@
 <th:block th:fragment="sidebar(active)">
     <meta http-equiv="Content-Type" content="text/html; charset=UTF-8"/>
     <link rel="stylesheet" href="./css/fragments.css" type="text/css" />
+
     <div class="navbar-header">
         <button type="button" class="navbar-toggle collapsed" data-toggle="collapse" data-target="#bs-example-navbar-collapse-1" aria-expanded="false">
             <span class="sr-only">Toggle navigation</span>
@@ -22,6 +23,7 @@
     <div class="collapse navbar-collapse" id="bs-example-navbar-collapse-1">
         <meta http-equiv="Content-Type" content="text/html; charset=UTF-8"/>
         <link rel="stylesheet" href="./css/fragments.css" type="text/css" />
+
 
         <ul class="nav navbar-nav">
             <!-- button to take user home -->
